rag-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.52.0
qdrant-client==1.12.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.3
markdown==3.7
beautifulsoup4==4.12.3
dashscope
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
cohere==5.5.0